More fixes so Tribler can be started from any path
