+++ Tribler/Main/tribler.py
@@ -1,22 +1,33 @@
+#!/usr/bin/env python
 import logging.config
 import sys
 import os
 
+
+from Tribler.Core.Utilities.install_dir import determine_install_dir
+# This should work for Linux, Windows and OSX
+TRIBLER_ROOT = determine_install_dir()
+LOGGER_CONF = os.path.join(TRIBLER_ROOT, "logger.conf")
+
+# Add TRIBLER_ROOT to the PYTHONPATH so imports work fine wherever Tribler is
+# running from.
+if TRIBLER_ROOT not in sys.path:
+    sys.path.append(TRIBLER_ROOT)
+
 try:
-    filepath = os.path.abspath(u"logger.conf")
-    logging.config.fileConfig(filepath)
+    logging.config.fileConfig(LOGGER_CONF)
 except Exception as e:
-    print >> sys.stderr, u"Unable to load logging config from '%s' file: %s" % (repr(filepath), repr(e))
+    print >> sys.stderr, u"Unable to load logging config from '%s' file: %s" % (repr(LOGGER_CONF), repr(e))
     print >> sys.stderr, u"Current working directory: %s" % repr(os.path.abspath(u'.'))
-    if not os.path.exists(filepath):
+    if not os.path.exists(LOGGER_CONF):
         print >> sys.stderr, "File doesn't exist"
-    elif not os.path.isfile(filepath):
+    elif not os.path.isfile(LOGGER_CONF):
         print >> sys.stderr, "It is not a file"
     else:
         print >> sys.stderr, "logger.conf file contents follow:"
         print >> sys.stderr, "8<--------------" * 5
 
-        with open(filepath, 'r') as conf:
+        with open(LOGGER_CONF, 'r') as conf:
             print >> sys.stderr, conf.read()
 
         print >> sys.stderr, "8<--------------" * 5
@@ -49,8 +60,17 @@ if sys.version_info[:2] != (2, 7):
     print >> sys.stderr, "Tribler needs python 2.7.X to run, current version: %s" % sys.version
     exit(1)
 
-if __name__ == '__main__':
+# Windows py2exe freezing tribler.py wreaks havoc with paths, imports, etc. So a
+# tribler_exe.py module has been created which only role is to find this module
+# (tribler.py), import __main__ and call it. The former is the one that will be
+# frozen by py2exe. Any other OSes can keep calling tribler.py just fine.
 
+def __main__():
+    """
+    Run Tribler and check if the reactor is dirty when shutting down.
+
+    This method is also called from tribler_exe.py.
+    """
     run()
     delayed_calls = reactor.getDelayedCalls()
     if delayed_calls:
@@ -59,3 +79,6 @@ if __name__ == '__main__':
             print >> sys.stderr, ">     %s" % dc
 
     stop_reactor()
+
+if __name__ == '__main__':
+    __main__()
